+++ lib/skeem/skm_binding.rb
@@ -1,17 +1,17 @@
 require_relative 'skm_element'
 
 module Skeem
-  
+
   # An identifier that is not a syntactic keyword can be used as a variable.
   # A variable may give a name to value that is bound (i.e. associated)
   # to that variable.
   class SkmBinding < SkmElement
     # @return [SkmIdentifier] The identifier that is bound the value.
     attr_reader(:variable)
-    
+
     # @return [SkmElement] The Skeem object that is associated with the variable.
     attr_reader(:value)
-    
+
     # Constructor
     # @param anIdentifier [SkmIdentifier] The variable name
     # @param aValue [SkmElement] The value to bind to the variable.
@@ -19,7 +19,7 @@ module Skeem
       @variable = anIdentifier
       @value = aValue
     end
-    
+
     def evaluate(aRuntime)
       name = variable.evaluate(aRuntime)
 
@@ -38,44 +38,12 @@ module Skeem
       else
         result = value.evaluate(aRuntime)
       end
-      
-=begin      
-      aRuntime.add_binding(var_key, self)
-      case expression
-        when SkmLambda
-          result = expression.evaluate(aRuntime)
-
-        when SkmVariableReference
-          other_key = expression.variable.evaluate(aRuntime)
-          if var_key.value != other_key.value
-            entry = aRuntime.fetch(other_key)
-            result = expression.evaluate(aRuntime)
-            if entry.kind_of?(Primitive::PrimitiveProcedure)
-              @expression = entry
-            elsif entry.kind_of?(SkmDefinition)
-              if entry.expression.kind_of?(SkmLambda)
-                @expression = entry.expression
-              end
-            end
-          else
-            # INFINITE LOOP DANGER: definition of 'x' has a reference to 'x'!
-            # Way out: the lookup for the reference should start from outer
-            # environment.
-            env = aRuntime.pop
-            @expression = expression.evaluate(aRuntime)
-            aRuntime.push(env)
-            result = expression
-          end
-        else
-          result = self
-      end
-=end
       binding_action(aRuntime, name, result)
       result
     end
-    
+
     protected
-    
+
     def binding_action(aRuntime, anIdentifier, anExpression)
       aRuntime.add_binding(anIdentifier, anExpression)
     end
@@ -87,17 +55,35 @@ module Skeem
       end
 
       result
-    end    
-    
+    end
+
   end # class
-  
+
 
   class SkmUpdateBinding < SkmBinding
-  
+
     protected
-    
+
     def binding_action(aRuntime, anIdentifier, anExpression)
       aRuntime.update_binding(anIdentifier, anExpression)
-    end    
+    end
   end # class
+  
+  class SkmDelayedUpdateBinding < SkmBinding
+    attr_reader :new_val
+    
+    def initialize(anIdentifier, aValue)
+      super(anIdentifier, aValue)
+    end
+    
+    def do_it!(aRuntime)
+      aRuntime.update_binding(variable, new_val)
+    end
+
+    protected
+
+    def binding_action(_runtime, _identifier, anExpression)
+      @new_val = anExpression
+    end
+  end # class  
 end # module